+++ src/Background/modules/launch.ts
@@ -2,8 +2,8 @@ import { thunkify } from 'ramda'
 import cfg from '../../config'
 import { Base, createBase, initLayoutInfo, RevertContainerID, selectSiteSettingsByFiltered } from '../../core/base'
 import { getControlWindowHeight } from '../../core/base/control-window-height'
+import { controlIsLaunched } from '../../core/control-window'
 import { calcControlWindowPos } from '../../core/layout/control-window'
-import { controlIsLaunched } from '../../hooks/useControlWindowExists'
 import { specifyFloorIdxBySearchText } from '../../hooks/useSearchForm'
 
 function getSiteSettings(base: Base, search_text: string) {
