+++ src/pages/Popup/floors/0-Search.tsx
@@ -2,7 +2,7 @@ import { thunkify } from 'ramda'
 import { Atomic } from 'vait'
 import React, { ReactNode, useEffect, useMemo, useState } from 'react'
 
-import { controlIsLaunched } from '../../../hooks/useControlWindowExists'
+import { controlIsLaunched } from '../../../core/control-window'
 import { sendMessage } from '../../../message'
 import launchControlWindow from '../../../Background/modules/launch'
